% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{add_all_columns}
\alias{add_all_columns}
\title{Add or remove columns of data}
\usage{
add_all_columns(data, vars)
}
\arguments{
\item{data}{tibble or dataframe containing the study data}

\item{vars}{additional data to be added}
}
\value{
list or tibble with additional data added
}
\description{
Add or remove columns of data as needed so that all sets
have the same columns. Used for adding extra columns for the trait, sites
and context tibbles
}
