% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process.R
\name{add_values_to_additional_trait_long}
\alias{add_values_to_additional_trait_long}
\title{Add values to an additional trait for datasets in long format}
\usage{
add_values_to_additional_trait_long(
  data,
  new_trait,
  traits_column,
  values_column,
  original_values,
  new_values
)
}
\arguments{
\item{data}{data frame, representing a specific dataset_id}

\item{new_trait}{name of the new variable being created, representing an additional trait in a long dataset}

\item{traits_column}{column containing the trait names}

\item{values_column}{column containing the trait values}

\item{original_values}{values of the original trait that need to be remapped to a different (new) trait}

\item{new_values}{values to be added to the new trait}
}
\value{
data frame containing additional rows of data for a new trait
}
\description{
Add values to an additional trait for datasets in long format
}
