% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{add_version_info}
\alias{add_version_info}
\title{Add version information to AusTraits}
\usage{
add_version_info(austraits, version, git_sha)
}
\arguments{
\item{austraits}{AusTraits database object}

\item{version}{version number}

\item{git_sha}{Git SHA}
}
\value{
AusTraits database object with version information added
}
\description{
Add version information to AusTraits
}
