% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{append_to_list}
\alias{append_to_list}
\title{Add an item to the end of a list}
\usage{
append_to_list(my_list, to_append)
}
\arguments{
\item{my_list}{a list}

\item{to_append}{a list}
}
\value{
a list merged with an added item at the end
}
\description{
Add an item to the end of a list
}
\examples{
 append_to_list(as.list(iris)[c(1,2)], as.list(iris)[c(3,4)])
}
