% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_utils.R
\name{as_link}
\alias{as_link}
\title{Generate hyperlink for markdown and html}
\usage{
as_link(link, text, type = "md")
}
\arguments{
\item{link}{character string for the url link}

\item{text}{character string for the text to display}

\item{type}{file type, default is markdown "md" otherwise html}
}
\value{
character string with the text and link formatted for md and html
}
\description{
Generate hyperlink for markdown and html files
}
\examples{
as_link("www.austraits.org", "austraits")
}
