% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{austraits_find_species}
\alias{austraits_find_species}
\title{Find list of all unique species within AusTraits}
\usage{
austraits_find_species(taxon_name, original_name = FALSE)
}
\arguments{
\item{taxon_name}{name of column which contains the cleaned species names}

\item{original_name}{name of column which contains original species names, default = FALSE}
}
\value{
list of all unique and distinct species names
}
\description{
Find list of all unique species within AusTraits
}
