% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{austraits_rebuild_remake_setup}
\alias{austraits_rebuild_remake_setup}
\title{Update the remake.yml file with new studies}
\usage{
austraits_rebuild_remake_setup()
}
\value{
Updated remake.yml file
}
\description{
`austraits_rebuild_remake_setup` rewrites the remake.yml file to include new
studies
}
