% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{austraits_rebuild_taxon_list}
\alias{austraits_rebuild_taxon_list}
\title{Builds list of potential species from the Australian Plant Census (APC) and 
Australian Plant Names Index (APNI)}
\usage{
austraits_rebuild_taxon_list()
}
\description{
Compiled list is saved at "config/taxon_list.csv". While this list is 
only an intermediate structure constructed entirely from 
the downloaded files, it saves us keeping copies of the entire 
lists (~8 vs 230Mb)
}
