% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{bib_print}
\alias{bib_print}
\title{Format BibEntry using RefManageR}
\usage{
bib_print(
  bib,
  .opts = list(first.inits = TRUE, max.names = 1000, style = "markdown")
)
}
\arguments{
\item{bib}{BibEntry object}

\item{.opts}{list of parameters for formatting style}
}
\value{
character string of formatted reference
}
\description{
Format BibEntry object according to desired style using RefManageR
}
