% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{build_add_version}
\alias{build_add_version}
\title{Add version information to AusTraits}
\usage{
build_add_version(austraits, version, git_sha)
}
\arguments{
\item{austraits}{AusTraits database object}

\item{version}{version number}

\item{git_sha}{Git SHA}
}
\value{
AusTraits database object with version information added
}
\description{
Add version information to AusTraits
}
