% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{build_combine}
\alias{build_combine}
\title{Combine all the AusTraits studies into the compiled AusTraits database}
\usage{
build_combine(..., d = list(...))
}
\arguments{
\item{...}{arguments passed to other functions}

\item{d}{list of all the AusTraits studies}
}
\value{
AusTraits compilation database as a large list
}
\description{
`build_combine` compiles all the loaded studies into a single AusTraits
database object as a large list
}
