% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_utils.R
\name{build_dataset_reports}
\alias{build_dataset_reports}
\title{Build reports for all studies}
\usage{
build_dataset_reports(dataset_ids = NULL, ...)
}
\arguments{
\item{dataset_ids}{names of studies/ datasets, default is NULL}

\item{...}{arguments passed to build_dataset_report()}
}
\value{
html files of the study report for all studies
}
\description{
Build reports for all studies using the `build_dataset_report()` function. This
function builds a study report for every study with a unique `dataset_id` that
has been loaded into AusTraits using `build_dataset_report()`. The reports are 
rendered as html files and saved in the "export/reports" folder.
}
