% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{build_find_taxon}
\alias{build_find_taxon}
\title{Find list of all unique taxa within compilation}
\usage{
build_find_taxon(taxon_name, austraits, original_name = FALSE)
}
\arguments{
\item{taxon_name}{name of column which contains the cleaned species names}

\item{austraits}{austraits compilation}

\item{original_name}{name of column which contains original species names, default = FALSE}
}
\value{
list of all unique and distinct species names
}
\description{
Find list of all unique taxa within compilation
}
