% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{build_setup_pipeline}
\alias{build_setup_pipeline}
\title{Update the remake.yml file with new studies}
\usage{
build_setup_pipeline(
  template = readLines(system.file("support", "remake.yml.whisker", package =
    "austraits.build")),
  path = "data",
  dataset_ids = dir(path)
)
}
\arguments{
\item{template}{template used to build}

\item{path}{path to folder with data}

\item{dataset_ids}{dataset_ids to icnlude. By default includes all.}
}
\value{
Updated remake.yml file
}
\description{
`build_setup_pipeline` rewrites the remake.yml file to include new
studies
}
