% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{check_all_values_in}
\alias{check_all_values_in}
\title{Check values in one vector against values in another vector}
\usage{
check_all_values_in(x, y, sep = " ")
}
\arguments{
\item{x}{vector}

\item{y}{vector}

\item{sep}{amount of space separating values to be split, default = " " (a single space)}
}
\value{
vector of logical values
}
\description{
Check values in one vector against values in another vector
}
