% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{check_missing_values}
\alias{check_missing_values}
\title{Check missing values from excluded data}
\usage{
check_missing_values(filename_data_csv, filename_metadata)
}
\arguments{
\item{filename_data_csv}{location of data.csv file}

\item{filename_metadata}{location of metadata.yml file}
}
\value{
tibble with excluded data including missing values.
}
\description{
`check_missing_values` shows missing values from the excluded_data table. 
Missing values are filtered out from the main austraits table and 
check_missing_values allows missing values to be checked for individual studies.
}
