% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{combine_austraits}
\alias{combine_austraits}
\title{Combine all the AusTraits studies into the compiled AusTraits database}
\usage{
combine_austraits(..., d = list(...), definitions)
}
\arguments{
\item{...}{arguments passed to other functions}

\item{d}{list of all the AusTraits studies}

\item{definitions}{definitions read in from the definitions.yml file in the config folder}
}
\value{
AusTraits compilation database as a large list
}
\description{
`combine_austraits` compiles all the loaded studies into a single AusTraits 
database object as a large list
}
