% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{combine_datasets}
\alias{combine_datasets}
\title{Combine all the AusTraits studies into the compiled AusTraits database}
\usage{
combine_datasets(..., d = list(...))
}
\arguments{
\item{...}{arguments passed to other functions}

\item{d}{list of all the AusTraits studies}
}
\value{
AusTraits compilation database as a large list
}
\description{
`combine_datasets` compiles all the loaded studies into a single AusTraits
database object as a large list
}
