% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process.R
\name{convert_01_ny}
\alias{convert_01_ny}
\title{Convert 0/1 to n/y}
\usage{
convert_01_ny(txt)
}
\arguments{
\item{txt}{vector of character strings}
}
\value{
vector of character strings
}
\description{
Convert binary 0/1 to n/y representing yes & no
}
\examples{
convert_01_ny(c(0,1,1,1,0))
}
