% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{convert_bib_to_list}
\alias{convert_bib_to_list}
\title{Convert BibEntry object to a list}
\usage{
convert_bib_to_list(bib)
}
\arguments{
\item{bib}{BibEntry object}
}
\value{
list
}
\description{
Convert BibEntry object to a list
}
