% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{convert_list_to_bib}
\alias{convert_list_to_bib}
\title{Convert a list of elements into a BibEntry object}
\usage{
convert_list_to_bib(ref)
}
\arguments{
\item{ref}{list of elements for a reference}
}
\value{
BibEntry object
}
\description{
Convert a list of elements into a BibEntry object
}
