% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process.R
\name{convert_month_range_string_to_binary}
\alias{convert_month_range_string_to_binary}
\title{Convert month range to 12 element binary string}
\usage{
convert_month_range_string_to_binary(str)
}
\arguments{
\item{str}{character string of month range using abbreviated month names. e.g. "Oct- Mar"}
}
\value{
a character string  length-12, e.g. "111000000111"
}
\description{
Converts flowering and fruiting month ranges to a string of 12 characters of binary data
consisting of 1 & 0 e.g. "11100000111"
}
