% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process.R
\name{convert_month_range_vec_to_binary}
\alias{convert_month_range_vec_to_binary}
\title{Convert vectors of month range to 12 element character strings of binary data}
\usage{
convert_month_range_vec_to_binary(vec)
}
\arguments{
\item{vec}{vector of month range using abbreviated month names (e.g. "Jan-Apr")}
}
\value{
character string of length 12 consisting of y & n, e.g. "yyyynnnnnnnn"
}
\description{
Converts vectors of month range type values (i.e. 'Jan-Apr') to character strings 
of length 12 consisting of y & n, e.g. "yyyynnnnnnnn"
}
