% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{convert_units}
\alias{convert_units}
\title{Convert units to desired type}
\usage{
convert_units(data, trait_definitions, unit_conversion_functions)
}
\arguments{
\item{data}{tibble or dataframe containing the study data}

\item{trait_definitions}{definitions read in from the traits.yml file in the config folder}

\item{unit_conversion_functions}{unit_conversions.csv file stored in the config folder}
}
\value{
tibble with converted units
}
\description{
Convert units to desired type
}
