% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{create_entity_id}
\alias{create_entity_id}
\title{Create entity id}
\usage{
create_entity_id(data)
}
\arguments{
\item{data}{the traits table at the point where this function is called}
}
\value{
character string
}
\description{
Creates 3-part entity id codes that combine a segment for species, population, 
and, when applicable, individual
This depends upon observation_id being established when the data.csv file is first read in
}
