% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{create_release}
\alias{create_release}
\title{Create release}
\usage{
create_release(austraits, v_prev = NULL)
}
\arguments{
\item{austraits}{AusTraits database object}

\item{v_prev}{specify whether to update the NEWS.md file, default = NULL}
}
\description{
Builds various files to go with the release of a new AusTraits version. Including
austraits.rds, a readme file, dictionary.rmd file, plain text files with csv, and
a NEWS.md file.
}
