% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{custom_manipulation}
\alias{custom_manipulation}
\title{Apply custom data manipulations}
\usage{
custom_manipulation(txt)
}
\arguments{
\item{txt}{character text within custom_R_code}
}
\value{
character text containing custom_R_code if custom_R_code is not empty,
otherwise no changes are made
}
\description{
Applies custom data manipulations if the metadata field custom_R_code is not empty
Otherwise no manipulations will be done.
}
