% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_utils.R
\name{dataset_generate_report}
\alias{dataset_generate_report}
\title{Build report for a specific study}
\usage{
dataset_generate_report(
  dataset_id,
  overwrite = FALSE,
  output_path = "export/reports",
  input_file = system.file("support", "report_dataset.Rmd", package =
    "austraits.build")
)
}
\arguments{
\item{dataset_id}{name of specific study/dataset}

\item{overwrite}{logical value to determine whether to overwrite existing report,
default = FALSE, if report exists already set to TRUE to overwrite}

\item{output_path}{location where rendered report will be saved}

\item{input_file}{report script (.Rmd) file to build study report}
}
\value{
html file of the rendered report located in the "export/reports" folder
}
\description{
Builds a report for a specified study using `dataset_id`. The information for 
the report is stored in an Rmd file and the final report is rendered as an html.  
The report is generated from the `report_study.Rmd` file in the scripts folder. 
Existing reports can be overwritten by setting overwrite = TRUE
}
