% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_utils.R
\name{dataset_generate_reports}
\alias{dataset_generate_reports}
\title{Build reports for all studies}
\usage{
dataset_generate_reports(dataset_ids = NULL, ...)
}
\arguments{
\item{dataset_ids}{names of studies/ datasets, default is NULL}

\item{...}{arguments passed to dataset_generate_report()}
}
\value{
html files of the study report for all studies
}
\description{
Build reports for all studies using the `dataset_generate_report()` function. This
function builds a study report for every study with a unique `dataset_id` that
has been loaded into AusTraits using `dataset_generate_report()`. The reports are 
rendered as html files and saved in the "export/reports" folder.
}
