% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{dataset_process}
\alias{dataset_process}
\title{Load Dataset}
\usage{
dataset_process(
  filename_data_raw,
  config_for_dataset,
  schema,
  filter_missing_values = TRUE
)
}
\arguments{
\item{filename_data_raw}{raw data.csv file for any given study}

\item{config_for_dataset}{config settings generated from `dataset_configure()`}

\item{schema}{schema for austraits.build}

\item{filter_missing_values}{default filters missing values from the excluded data table. 
Change to false to see the rows with missing values.}
}
\value{
list, AusTraits database object
}
\description{
dataset_process is used to load individual studies using the config file generated
from `dataset_configure()`. `dataset_configure` and `dataset_process` are applied to every
study in the remake.yml file
}
\examples{
\dontrun{
dataset_process("data/Falster_2003/data.csv", dataset_configure("data/Falster_2003/metadata.yml",
read_yaml("config/traits.yml"), get_unit_conversions("config/unit_conversions.csv")),
get_schema())
}
}
