% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{dataset_report}
\alias{dataset_report}
\title{Build reports for listed datasets}
\usage{
dataset_report(
  dataset_id,
  austraits,
  overwrite = FALSE,
  output_path = "export/reports",
  input_file = system.file("support", "report_dataset.Rmd", package =
    "austraits.build"),
  quiet = TRUE,
  keep = FALSE
)
}
\arguments{
\item{dataset_id}{name of specific study/dataset}

\item{austraits}{compiled austraits database}

\item{overwrite}{logical value to determine whether to overwrite existing report,}

\item{output_path}{location where rendered report will be saved}

\item{input_file}{report script (.Rmd) file to build study report}

\item{quiet}{An option to suppress printing during rendering from knitr, pandoc command line and others.}

\item{keep}{keep intermediate Rmd file used?}
}
\value{
html file of the rendered report located in the specified output folder.
}
\description{
Builds a detailed report for every dataset with a unique `dataset_id`, based on the template Rmd file provided.  The reports are
rendered as html files and saved in the specified output folder.
}
