% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_datasets.R
\name{dataset_test_setup}
\alias{dataset_test_setup}
\title{Test whether specified dataset_id has the correct setup}
\usage{
dataset_test_setup(
  dataset_ids,
  path_config = "config",
  path_data = "data",
  reporter = testthat::default_reporter()
)
}
\arguments{
\item{dataset_ids}{vector of dataset_id for sources to be tested}

\item{path_config}{path to folder containing configuration files}

\item{path_data}{path to folder containing data files}

\item{reporter}{testthat reporter to use to summarise output}
}
\description{
Run tests to ensure that specified dataset_id has the correct setup
}
