% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdata.R
\name{dataset_test_worker}
\alias{dataset_test_worker}
\title{Test whether specified dataset_id has the correct setup}
\usage{
dataset_test_worker(
  test_dataset_ids,
  path_config = "config",
  path_data = "data",
  schema = get_schema(),
  definitions = get_schema(file.path(path_config, "traits.yml"), I("traits"))
)
}
\arguments{
\item{test_dataset_ids}{vector of dataset_id for sources to be tested}

\item{path_config}{path to folder containing configuration files}

\item{path_data}{path to folder containing data files}

\item{schema}{data schema}

\item{definitions}{trait defininitons}
}
\description{
Run tests to ensure that specified dataset_id has the correct setup
}
