% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{df_to_list}
\alias{df_to_list}
\title{Convert dataframe to list}
\usage{
df_to_list(df)
}
\arguments{
\item{df}{a dataframe}
}
\value{
a (yaml) list
}
\description{
Convert a dataframe to a named list, useful when converting to yaml
}
\examples{
df_to_list(iris)
}
