% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{export_version_plaintext}
\alias{export_version_plaintext}
\title{Export AusTraits version as plain text}
\usage{
export_version_plaintext(austraits, path)
}
\arguments{
\item{austraits}{AusTraits database object}

\item{path}{pathway to save file}
}
\value{
csv files of tibbles containing traits, sites, contexts, methods, excluded_data,
taxonomic updates, taxa, contributors
}
\description{
Export AusTraits version as plain text
}
