% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{find_names_distance_to_neighbours}
\alias{find_names_distance_to_neighbours}
\title{Find the distance for nearby species (needs review)}
\usage{
find_names_distance_to_neighbours(taxon_name, dist = 5)
}
\arguments{
\item{taxon_name}{vector of species names}

\item{dist}{numerical value for distance, default = 5}
}
\value{
a vector of distances between species
}
\description{
Find the distance for nearby species (needs review)
}
