% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process.R
\name{format_flowering_months}
\alias{format_flowering_months}
\title{Create flowering times from start to end pair}
\usage{
format_flowering_months(start, end)
}
\arguments{
\item{start}{numeric value between 1 and 12 representing earliest flowering month}

\item{end}{numeric value between 1 and 12 representing latest flowering month}
}
\value{
a 12 element character string consisting of y & n, e.g. "yyyynnnnnnnn"
}
\description{
A common processing pattern when creating flowering times
We have an integer for start and end month
Converts these into binary vector
}
