% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process.R
\name{format_min_max_as_range}
\alias{format_min_max_as_range}
\title{Merge minimum and maximum trait values into a single range column.}
\usage{
format_min_max_as_range(
  data,
  min_column,
  max_column,
  range_column,
  column_value_type
)
}
\arguments{
\item{data}{name of the dataset being mutated}

\item{min_column}{name of the variable with the minimum trait values in the wide datas file}

\item{max_column}{name of the variable with the maximum trait values in the wide datas file}

\item{range_column}{name of a variable being created that combines the minimum and maximum values}

\item{column_value_type}{name of a variable being created that indicates the value type of the associated column}
}
\value{
data frame with a two new columns containing manipulated trait data
}
\description{
This is an appropriate function to use for recombining range values extracted from floras and taxonomic treatments that were previously separated in AusTraits
}
\examples{
\dontrun{
data <- read_csv("data/Chinnock_2007/data.csv")
data \%>\%
 format_min_max_as_range("min leaf length (mm)", "max leaf length (mm)", "leaf_length_range", "leaf_length_value_type") \%>\%
 format_min_max_as_range("min leaf width (mm)", "max leaf width (mm)", "leaf_width_range", "leaf_width_value_type") \%>\%
 format_min_max_as_range("seed length min (mm)", "seed length max (mm)", "seed_length_range", "seed_length_value_type")
}
}
