% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process.R
\name{get_month}
\alias{get_month}
\title{Return month for given indices}
\usage{
get_month(i)
}
\arguments{
\item{i}{vector with indices/integers representing months}
}
\value{
vector containing abbreviated months
}
\description{
Returns abbreviated month for given indices/ integers,
if needed, i is coerced to integer
warnings suppressed because as.integer gives a warning for NAs
}
\examples{
get_month(c(1,1,2,3,3,3,4,5,6))
}
