% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_utils.R
\name{get_SHA_link}
\alias{get_SHA_link}
\title{Get SHA link from Github}
\usage{
get_SHA_link(...)
}
\arguments{
\item{...}{arguments passed to the get_SHA()}
}
\value{
SHA link to a github commit as a character string formatted using markdown syntax
}
\description{
Get SHA link using the get_SHA() function. The link generated leads to the latest
commit for the Github repository. SHA is the abbreviated SHA-1 40 digit
hexadecimal number which Github uses to track commits and changes made to a repository.
}
