% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_schema.R
\name{get_version}
\alias{get_version}
\title{Retrieve version for compilation from definitions}
\usage{
get_version(path)
}
\arguments{
\item{path}{path to schema file.}
}
\value{
a string
}
\description{
Retrieve version for compilation from definitions
}
