% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/austraits.R
\name{label_suspected_duplicates}
\alias{label_suspected_duplicates}
\title{Check possible duplicates across studies}
\usage{
label_suspected_duplicates(austraits_traits, priority_sources = NULL)
}
\arguments{
\item{austraits_traits}{AusTraits traits data frame}

\item{priority_sources}{default is NULL, a vector of studies to check against 
for duplicates}
}
\value{
data frame containing potential duplicated data
}
\description{
`label_suspected_duplicates` checks across studies to identify possible
duplicates
}
