% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{list_to_df}
\alias{list_to_df}
\title{Convert a list of lists to dataframe}
\usage{
list_to_df(my_list, as_character = TRUE, on_empty = NA)
}
\arguments{
\item{my_list}{a list of lists to dataframe}

\item{as_character}{logical:  indicating whether the values are read as character}

\item{on_empty}{value to return if my_list is NULL, NA or is length == 0, default = NA}
}
\description{
Convert a list of lists to dataframe requires that every list have same named elements
}
\examples{
list_to_df(df_to_list(iris))
}
