% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{load_dataset}
\alias{load_dataset}
\title{Load Dataset}
\usage{
load_dataset(
  filename_data_raw,
  config_for_dataset,
  schema,
  filter_missing_values = TRUE
)
}
\arguments{
\item{filename_data_raw}{raw data.csv file for any given study}

\item{config_for_dataset}{config settings generated from `subset_config()`}

\item{schema}{schema for austraits.build}

\item{filter_missing_values}{default filters missing values from the excluded data table. 
Change to false to see the rows with missing values.}
}
\value{
list, AusTraits database object
}
\description{
load_dataset is used to load individual studies using the config file generated
from `subset_config()`. `subset_config` and `load_dataset` are applied to every
study in the remake.yml file
}
\examples{
\dontrun{
load_dataset("data/Falster_2003/data.csv", subset_config("data/Falster_2003/metadata.yml",
read_yaml("config/traits.yml"), make_unit_conversion_functions("config/unit_conversions.csv")),
load_schema())
}
}
