% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_schema.R
\name{load_schema}
\alias{load_schema}
\title{Load schema for an austraits.build data compilation (excluding traits)}
\usage{
load_schema(
  path = system.file("support", "austraits.build_schema.yml", package =
    "austraits.build"),
  subsection = NULL
)
}
\arguments{
\item{path}{path to schema file. By default loads version included with the package}

\item{subsection}{section to load}
}
\value{
a list
}
\description{
Load schema for an austraits.build data compilation (excluding traits)
}
\examples{
{

schema <- load_schema()
}
}
