% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{load_study}
\alias{load_study}
\title{Load Study}
\usage{
load_study(filename_data_raw, config_for_dataset)
}
\arguments{
\item{filename_data_raw}{raw data.csv file for any given study}

\item{config_for_dataset}{config settings generated from `subset_config()`}
}
\value{
list, AusTraits database object
}
\description{
load_study is used to load individual studies using the config file generated
from `subset_config()`. `subset_config` and `load_study` are applied to every
study in the remake.yml file
}
\examples{
\dontrun{
load_study("data/Falster_2003/data.csv", subset_config("data/Falster_2003/metadata.yml",
read_yaml("config/definitions.yml"), make_unit_conversion_functions("config/unit_conversions.csv"))
}
}
