% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{load_taxonomic_resources}
\alias{load_taxonomic_resources}
\title{Load taxonomic resources from the APC and APNI}
\usage{
load_taxonomic_resources(
  path_apc = "config/NSL/APC-taxon-2020-05-14-1332.csv",
  path_apni = "config/NSL/APNI-names-2020-05-14-1341.csv"
)
}
\arguments{
\item{path_apc}{location of downloaded APC taxon file}

\item{path_apni}{location of downloaded APNI name file}
}
\description{
Load taxonomic resources from the Australian Plant Census and the Australian 
Plant Name Index. Taxonomic resources are stored as csv files in the NSL folder
}
