% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{make_unit_conversion_functions}
\alias{make_unit_conversion_functions}
\title{Make unit conversion functions}
\usage{
make_unit_conversion_functions(filename)
}
\arguments{
\item{filename}{name of file containing unit conversions}
}
\value{
list of conversion functions
}
\description{
Make unit conversion functions
}
\examples{
\dontrun{
make_unit_conversion_functions("config/unit_conversions.csv")
}
}
