% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{metadata_add_contexts}
\alias{metadata_add_contexts}
\title{For specified `dataset_id` import context data from a dataframe}
\usage{
metadata_add_contexts(dataset_id, overwrite = FALSE)
}
\arguments{
\item{dataset_id}{identifier for a particular study in the AusTraits database}

\item{overwrite}{Overwrite existing information}
}
\description{
This functions asks users which columns in the dataframe they would like to keep
and records this appropriately in the metadata. The input data is assumed to be 
in wide format.
The output may require additional manual editing.
}
