% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{metadata_add_source_bibtex}
\alias{metadata_add_source_bibtex}
\title{Adds citation details to a metadata file for given study}
\usage{
metadata_add_source_bibtex(
  dataset_id,
  file,
  type = "primary",
  key = dataset_id,
  drop = c("dateobj", "month")
)
}
\arguments{
\item{dataset_id}{identifier for a particular study in the AusTraits database}

\item{file}{Name of file where reference is saved}

\item{type}{Type of references: `primary` or `secondary`}

\item{key}{The bibtex key to be used. By default set to `dataset_id`}

\item{drop}{Variables to ignore}
}
\value{
yml file with citation details added
}
\description{
Adds citation details to a metadata file for given study
}
