% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{metadata_add_substitution}
\alias{metadata_add_substitution}
\title{Add a categorical trait value substitution into a metadata file for a dataset_id}
\usage{
metadata_add_substitution(dataset_id, trait_name, find, replace)
}
\arguments{
\item{dataset_id}{identifier for a particular study in the AusTraits database}

\item{trait_name}{the AusTraits defined name for a particular trait}

\item{find}{trait value in the original data.csv file}

\item{replace}{trait value supported by AusTraits}
}
\value{
yaml file with a substitution added
}
\description{
`metadata_add_substitution` is used to align the categorical trait values used
by a contributor to the categorical values supported by AusTraits. These values
are defined in the `traits.yml` file
}
