% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{metadata_add_taxonomic_change}
\alias{metadata_add_taxonomic_change}
\title{Add a taxonomic change into the metadata yaml file for a dataset_id}
\usage{
metadata_add_taxonomic_change(
  dataset_id,
  find,
  replace,
  reason,
  taxonomic_resolution
)
}
\arguments{
\item{dataset_id}{identifier for a particular study in the AusTraits database}

\item{find}{original name used by the contributor}

\item{replace}{taxonomic name accepted by APC or APNI}

\item{reason}{reason for taxonomic change}
}
\value{
yml file with taxonomic change added
}
\description{
Add a single taxonomic change into the metadata yaml file for a specific study
}
