% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{metadata_check_taxa}
\alias{metadata_check_taxa}
\title{Check taxa against list of known species}
\usage{
metadata_check_taxa(
  dataset_id,
  max_distance_abs = 3,
  max_distance_rel = 0.2,
  try_outside_guesses = FALSE,
  author = git2r::config()$global$user.name
)
}
\arguments{
\item{dataset_id}{identifier for a particular study in the AusTraits database}

\item{max_distance_abs}{numerical value for absolute max distance, default = 3}

\item{max_distance_rel}{numerical value for relative max distance, default = 0.2}

\item{try_outside_guesses}{logical value, default = FALSE}

\item{author}{name of author}
}
\description{
Checks all taxa within against our list of known species
If not found, and update=TRUE, checks the unknown species against
}
