% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{metadata_create_template}
\alias{metadata_create_template}
\title{Create a template of file `metadata.yml` for specified `dataset_id`}
\usage{
metadata_create_template(
  dataset_id,
  path = file.path("data", dataset_id),
  skip_manual = FALSE
)
}
\arguments{
\item{dataset_id}{identifier for a particular study in the AusTraits database}

\item{path}{location of file where output is saved}

\item{skip_manual}{allows skipping of manual selection of variables, default = FALSE}
}
\value{
a yml file template for metadata
}
\description{
Includes place-holders for major sections of the metadata
}
