% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{metadata_find_taxonomic_change}
\alias{metadata_find_taxonomic_change}
\title{Find taxonomic changes within the metadata yml files}
\usage{
metadata_find_taxonomic_change(find, replace = NULL, studies = NULL)
}
\arguments{
\item{find}{name of original species}

\item{replace}{name of replacement species, default = NULL}

\item{studies}{name of studies, default = NULL}
}
\description{
Find taxonomic changes within the metadata yml files
}
