% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{metadata_path_dataset_id}
\alias{metadata_path_dataset_id}
\title{Path to the `metadata.yml` file for specified `dataset_id`}
\usage{
metadata_path_dataset_id(dataset_id)
}
\arguments{
\item{dataset_id}{identifier for a particular study in the AusTraits database}
}
\value{
A string
}
\description{
Path to the `metadata.yml` file for specified `dataset_id`
}
