% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{metadata_user_select_column}
\alias{metadata_user_select_column}
\title{Select column by user}
\usage{
metadata_user_select_column(column, choices)
}
\arguments{
\item{column}{name of the variable of interest}

\item{choices}{the options that can be selected from}
}
\description{
`metadata_user_select_column` is used to select which columns in a dataframe/ tibble 
corresponds to the variable of interest. It is used compile the metadata yaml
file by prompting the user to choose the relevant columns. It is used in 
`metadata_add_locations` and `metadata_add_contexts` and `metadata_create_template`
}
