% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{metadata_write_dataset_id}
\alias{metadata_write_dataset_id}
\title{Write the YAML representation of metadata.yml for specified `dataset_id` to
file \code{data/dataset_id/metadata.yml}}
\usage{
metadata_write_dataset_id(metadata, dataset_id)
}
\arguments{
\item{metadata}{metadata file}

\item{dataset_id}{identifier for a particular study in the AusTraits database}
}
\value{
a yml file
}
\description{
Write the YAML representation of metadata.yml for specified `dataset_id` to
file \code{data/dataset_id/metadata.yml}
}
