% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process.R
\name{move_values_to_new_trait}
\alias{move_values_to_new_trait}
\title{Move select trait values from a pre-existing column (trait_name) to a new column (new trait_name)}
\usage{
move_values_to_new_trait(
  data,
  original_trait,
  new_trait,
  original_values,
  values_for_new_trait,
  values_to_keep
)
}
\arguments{
\item{data}{data frame, representing a specific dataset_id}

\item{original_trait}{name of the variable in the original data file, representing a trait in a wide dataset}

\item{new_trait}{name of the new variable being created, representing an additional trait in a wide dataset}

\item{original_values}{values of the original trait that need to be remapped to a different (new) trait}

\item{value_for_new_trait}{the appropriate value of the new trait; this may be identical to the original values, or may be a slightly different word/syntax}

\item{value_to_keep}{the appropriate value to retain for the old trait; this may be identical to the original values or may be NA}
}
\value{

}
\description{
Move select trait values from a pre-existing column (trait_name) to a new column (new trait_name)
}
\examples{
data <- read_csv(data/"Hughes_1992/data.csv")
data \%>\% move_values_to_new_trait(data, "growth form", "root_structure", "Saprophyte", "saprophyte") -> data
}
