% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process.R
\name{move_values_to_new_trait_long}
\alias{move_values_to_new_trait_long}
\title{Move values in new trait in long format}
\usage{
move_values_to_new_trait_long(
  data,
  original_trait,
  new_trait,
  traits_column,
  values_column,
  original_values
)
}
\arguments{
\item{data}{data frame, representing a specific dataset_id}

\item{original_trait}{name of the variable in the original data file, representing a trait in a long dataset}

\item{new_trait}{name of the new variable being created, representing an additional trait in a long dataset}

\item{traits_column}{column containing the trait names}

\item{values_column}{column containing the trait values}

\item{original_values}{values of the original trait that need to be remapped to a different (new) trait}
}
\value{
data frame containing additional trait names in the traits column
}
\description{
Move values in new trait in long format
}
