% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_utils.R
\name{my_kable_styling_markdown}
\alias{my_kable_styling_markdown}
\title{Format table with kable and default styling for markdown}
\usage{
my_kable_styling_markdown(...)
}
\arguments{
\item{...}{arguments passed to `kableExtra::kable()`}
}
\description{
Format table with kable and default styling for markdown
}
