% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_utils.R
\name{my_kable_styling_pdf}
\alias{my_kable_styling_pdf}
\title{Format table with kable and default styling for pdf}
\usage{
my_kable_styling_pdf(...)
}
\arguments{
\item{...}{arguments passed to `kableExtra::kable()`}
}
\description{
Format table with kable and default styling for pdf
}
