% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notetaker.R
\name{notetaker_add_note}
\alias{notetaker_add_note}
\title{Add a note to the note recorder as a new row}
\usage{
notetaker_add_note(notes, new_note)
}
\arguments{
\item{notes}{object containing the report notes}

\item{new_note}{vector of character notes to be added to existing notes}
}
\value{
A tibble with additional notes added
}
\description{
Add a note to the note recorder as a new row
}
