% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notetaker.R
\name{notetaker_get_note}
\alias{notetaker_get_note}
\title{Return a specific row from notes}
\usage{
notetaker_get_note(notes, i = nrow(notes))
}
\arguments{
\item{notes}{object containing the report notes}

\item{i}{numerical; row number for corresponding note, default is nrow(notes)}
}
\value{
a single row from a tibble
}
\description{
Returns a specific row from notes specified by i. Default is nrow(notes) which 
returns the last note
}
