% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notetaker.R
\name{notetaker_print_all}
\alias{notetaker_print_all}
\title{Print all notes}
\usage{
notetaker_print_all(notes, ..., numbered = TRUE)
}
\arguments{
\item{notes}{object containing the report notes}

\item{...}{arguments passed to other functions}

\item{numbered}{logical default is TRUE}
}
\value{
character string containing the notes
}
\description{
Print all notes
}
