% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notetaker.R
\name{notetaker_print_note}
\alias{notetaker_print_note}
\title{Print note (needs review?)}
\usage{
notetaker_print_note(
  note,
  as_anchor = FALSE,
  anchor_text = "",
  link_text = "link"
)
}
\arguments{
\item{note}{object containing the report notes}

\item{as_anchor}{logical default is FALSE}

\item{anchor_text}{character string, default is ""}

\item{link_text}{character string, default is "link"}
}
\value{
character string containing the notes
}
\description{
Print note (needs review?)
}
