% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notetaker.R
\name{notetaker_print_notes}
\alias{notetaker_print_notes}
\title{Print a specific row from notes}
\usage{
notetaker_print_notes(notes, i = nrow(notes), ...)
}
\arguments{
\item{notes}{object containing the report notes}

\item{i}{specify the row which contains the note to be returned}

\item{...}{arguments passed to notetaker_print_note()}
}
\value{
character string containing the notes
}
\description{
Prints a specific row from notes specified by i
}
