% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notetaker.R
\name{notetaker_start}
\alias{notetaker_start}
\title{Start note recorder (needs review?)}
\usage{
notetaker_start()
}
\value{
A tibble where notes are recorded
}
\description{
Note recorder used in report_study.Rmd file to initiate note recorder
}
