% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{null_as}
\alias{null_as}
\title{Convert NULL values to a different value}
\usage{
null_as(x, val = NA)
}
\arguments{
\item{x}{a vector containing null values}

\item{val}{specify what the null value should be returned as, default is NA}
}
\value{
a vector with null values replaced
}
\description{
null_as converts NULL values in a vector to a different value. Default is
converting NuLL to NA
}
\examples{
null_as(NULL)
}
