% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{parse_data}
\alias{parse_data}
\title{Process a single dataset}
\usage{
parse_data(data, dataset_id, metadata)
}
\arguments{
\item{data}{tibble or dataframe containing the study data}

\item{dataset_id}{identifier for a particular study in the AusTraits database}

\item{metadata}{yaml file with metadata}
}
\value{
tibble
}
\description{
Processes a single dataset using the original dataframe and the associated 
metadata.yml file based on `dataset_id`. Builds a tibble with the study data 
and adds the corrected trait names, assigns a unique observation_id and implements
any categorical trait value substituions. `parse data` is used in the core 
workflow pipeline (such as in `load study`).
}
