% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_notetaker.R
\name{print_all_notes}
\alias{print_all_notes}
\title{Print all notes}
\usage{
print_all_notes(notes, ..., numbered = TRUE)
}
\arguments{
\item{notes}{object containing the report notes}

\item{...}{arguments passed to other functions}

\item{numbered}{logical default is TRUE}
}
\value{
character string containing the notes
}
\description{
Print all notes
}
