% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_add_all_columns}
\alias{process_add_all_columns}
\title{Add or remove columns of data}
\usage{
process_add_all_columns(data, vars, add_error_column = TRUE)
}
\arguments{
\item{data}{dataframe containing study data read in as a csv file}

\item{vars}{vector of variable columns names to be included in the final formatted tibble}

\item{add_error_column}{adds an extra column called error if TRUE}
}
\value{
tibble with the correct selection of columns including an error column
}
\description{
Add or remove columns of data as needed so that all datasets
have the same columns. Also adds in an error column.
}
