% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_create_observation_id}
\alias{process_create_observation_id}
\title{Create entity id}
\usage{
process_create_observation_id(data)
}
\arguments{
\item{data}{the traits table at the point where this function is called}
}
\value{
character string
}
\description{
Creates 3-part entity id codes that combine a segment for species, population, 
and, when applicable, individual
This depends upon a parsing_id being established when the data.csv file is first read in
}
