% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_custom_code}
\alias{process_custom_code}
\title{Apply custom data manipulations}
\usage{
process_custom_code(txt)
}
\arguments{
\item{txt}{character text within custom_R_code of a metadata.yml file}
}
\value{
character text containing custom_R_code if custom_R_code is not empty,
otherwise no changes are made
}
\description{
Applies custom data manipulations if the metadata field custom_R_code is not empty
Otherwise no manipulations will be done by applying the `identity` function.
The code custom_R_code assumes a single input.
}
