% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_flag_excluded_observations}
\alias{process_flag_excluded_observations}
\title{Flag any excluded observations}
\usage{
process_flag_excluded_observations(data, metadata)
}
\arguments{
\item{data}{tibble or dataframe containing the study data}

\item{metadata}{metadata yaml file for any given study}
}
\value{
dataframe with flagged excluded observations if there are any
}
\description{
Checks the metadata yaml file for any excluded observations. If there are none,
returns the original data. If there are excluded observations returns the mutated data
with excluded observations flagged in a new column
}
