% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_flag_unsupported_traits}
\alias{process_flag_unsupported_traits}
\title{Flag any unrecognised traits}
\usage{
process_flag_unsupported_traits(data, definitions)
}
\arguments{
\item{data}{tibble or dataframe containing the study data}

\item{definitions}{definitions read in from the traits.yml file in the config folder}
}
\value{
tibble with unrecognised traits flagged as "Unsupported trait" in the "error" column
}
\description{
Flag any unrecognised traits, as defined in the traits.yml file
}
