% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_format_contexts}
\alias{process_format_contexts}
\title{Format context data from list to tibble}
\usage{
process_format_contexts(my_list, dataset_id)
}
\arguments{
\item{my_list}{list of input information}

\item{dataset_id}{XXX}
}
\value{
tibble with context details if available
}
\description{
Format context data read in from the metadata.yml file. Converts from list to tibble.
}
\examples{
\dontrun{
process_format_contexts(read_metadata("data/Apgaua_2017/metadata.yml")$context)
}
}
