% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_format_locations}
\alias{process_format_locations}
\title{Format location data from list to tibble}
\usage{
process_format_locations(my_list, dataset_id, schema)
}
\arguments{
\item{my_list}{list of input information}

\item{dataset_id}{identifier for a particular study in the AusTraits database}

\item{schema}{XXX}
}
\value{
tibble with location details if available
}
\description{
Format location data read in from the metadata.yml file. Converts from list to tibble.
}
\examples{
\dontrun{
process_format_locations(read_metadata("data/Falster_2003/metadata.yml")$locations, "Falster_2003")
}
}
