% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_format_sites}
\alias{process_format_sites}
\title{Format site and context data from list to tibble}
\usage{
process_format_sites(my_list, dataset_id, context = FALSE)
}
\arguments{
\item{my_list}{list of input information}

\item{dataset_id}{identifier for a particular study in the AusTraits database}

\item{context}{logical, adds context information if available, default = FALSE}
}
\value{
tibble with site and context details if available
}
\description{
Format site data read in from the metadata.yml file. Converts from list to tibble.
Includes context information if available by specifying context = TRUE
}
\examples{
\dontrun{
process_format_sites(yaml::read_yaml("data/Falster_2003/metadata.yml")$sites, "Falster_2003")
process_format_sites(yaml::read_yaml("data/Apgaua_2017/metadata.yml")$context,
"Apgaua_2017", context = TRUE)
}
}
