% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{process_rename_columns}
\alias{process_rename_columns}
\title{Rename column names}
\usage{
process_rename_columns(obj, from, to)
}
\arguments{
\item{obj}{tibble or data frame with 1 or more columns}

\item{from}{character string of the initial column name}

\item{to}{character string of the new column name}
}
\value{
a tibble with new column names
}
\description{
`process_rename_columns` renames column names within a tibble or data frame by matching
the names given by the argument `from`. The new column names are renamed using the 
names given by the argument `to`.
}
