% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_standardise_names}
\alias{process_standardise_names}
\title{Standardise species names}
\usage{
process_standardise_names(x)
}
\arguments{
\item{x}{vector, dataframe or list containing original species names}
}
\value{
vector with standardised species names
}
\description{
Enforces some standards on species names
}
