% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_taxonomic_updates}
\alias{process_taxonomic_updates}
\title{Apply taxonomic updates}
\usage{
process_taxonomic_updates(data, metadata)
}
\arguments{
\item{data}{tibble or dataframe containing the study data}

\item{metadata}{yaml file containing the metadata}
}
\value{
tibble with the taxonomic updates applied
}
\description{
Applies taxonomic updates to the study data from the metadata.yml file
}
