% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_unit_conversion_name}
\alias{process_unit_conversion_name}
\title{Generate unit conversion name}
\usage{
process_unit_conversion_name(from, to)
}
\arguments{
\item{from}{character of original units}

\item{to}{character of units to be converted to}
}
\value{
character string containing the name what units are being converted to
}
\description{
creates the unit conversion name based on the original units and the units to
converted to
}
