% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_notetaker.R
\name{random_string}
\alias{random_string}
\title{Create a string of random letters}
\usage{
random_string(n = 8)
}
\arguments{
\item{n}{numerical integer, default is 8}
}
\value{
character string with 8 letters
}
\description{
Creates a string of random letters with 8 characters as the default,
useful for defining unique hyperlinks
}
