% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_metadata}
\alias{read_metadata}
\title{Read in a metadata.yml file for a study}
\usage{
read_metadata(path)
}
\arguments{
\item{path}{location of the metadata file}
}
\description{
Read in a metadata.yml file for a study
}
