% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_metadata_dataset}
\alias{read_metadata_dataset}
\title{Read the `metadata.yml` file for specified `dataset_id`}
\usage{
read_metadata_dataset(dataset_id)
}
\arguments{
\item{dataset_id}{identifier for a particular study in the AusTraits database}
}
\value{
A list with contents of metadata for specified `dataset_id`
}
\description{
Read the `metadata.yml` file for specified `dataset_id`
}
