% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/austraits.R
\name{remove_suspected_duplicates}
\alias{remove_suspected_duplicates}
\title{Move suspected duplicates to excluded data}
\usage{
remove_suspected_duplicates(austraits, priority_sources = NULL)
}
\arguments{
\item{austraits}{AusTraits data object}

\item{priority_sources}{default = NULL. A vector of studies to check against
for possible duplicates}
}
\value{
AusTraits data object
}
\description{
`remove_suspected_duplicates` moves suspected duplicates from the 'traits' table 
to excluded_data.
}
