% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{rename_columns}
\alias{rename_columns}
\title{Rename columns}
\usage{
rename_columns(obj, from, to)
}
\arguments{
\item{obj}{a  tibble with multiple columns}

\item{from}{a vector of the initial column names}

\item{to}{a vector of the new column names}
}
\value{
a  tibble with new column names
}
\description{
Rename columns
}
