% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process.R
\name{replace_duplicates_with_NA}
\alias{replace_duplicates_with_NA}
\title{Replace duplicate values with NA}
\usage{
replace_duplicates_with_NA(x)
}
\arguments{
\item{x}{vector containing values}
}
\value{
vector with duplicate values as NA
}
\description{
replace_duplicates_with_NA replaces duplicate values in x with NA.
}
