% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{run_benchmark}
\alias{run_benchmark}
\title{Check performance on current system using package bench}
\usage{
run_benchmark()
}
\value{
A dataframe of results
}
\description{
Check performance on current system using package bench
}
