% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process.R
\name{separate_range}
\alias{separate_range}
\title{Separate cells with a range to min and max}
\usage{
separate_range(data, x, y1, y2, sep = "-", remove = TRUE)
}
\arguments{
\item{data}{data frame}

\item{x}{name of variable containing range}

\item{y1}{name of variable to hold minimum}

\item{y2}{name of variable to hold maximum}

\item{sep}{separator, by default "-"}

\item{remove}{logical, default = TRUE}
}
\value{
modified data frame
}
\description{
Separate values cells with a range into columns with minimum and maximum
}
