% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{split_then_sort}
\alias{split_then_sort}
\title{Split and sort cells with multiple values}
\usage{
split_then_sort(x, sep = " ")
}
\arguments{
\item{x}{an individual cell with multiple values}

\item{sep}{a separator, a whitespace is the default}
}
\value{
a vector of alphabetically sorted records
}
\description{
split_then_sort: For a vector x in which individual cell may have
 multiple values (separated by 'sep'), sort records within each cell alphabetically
}
\examples{
split_then_sort("z y x")
}
