% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{standardise_names}
\alias{standardise_names}
\title{Standardise species names}
\usage{
standardise_names(x)
}
\arguments{
\item{x}{vector, dataframe or list containing original species names}
}
\value{
vector with standardised species names
}
\description{
Standardise species names
}
