% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_notetaker.R
\name{start_notetaker}
\alias{start_notetaker}
\title{Start note recorder (needs review?)}
\usage{
start_notetaker()
}
\value{
A tibble where notes are recorded
}
\description{
Note recorder used in report_study.Rmd file to initiate note recorder
}
