% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{subset_config}
\alias{subset_config}
\title{Configure AusTraits database object}
\usage{
subset_config(filename_metadata, definitions, unit_conversion_functions)
}
\arguments{
\item{filename_metadata}{metadata yaml file for a given study}

\item{definitions}{definitions read in from the traits.yml}

\item{unit_conversion_functions}{unit_conversion.csv file read in from the config folder}
}
\value{
list with dataset_id, metadata, definitions and unit_conversion_functions
}
\description{
Creates the config object which gets passed onto `load_dataset`. The config list contains
the subset of definitions and unit conversions for those traits for a each study.
`subset_config` is used in the remake::make process to configure individual studies mapping the
individual traits found in that study along with any relevant unit conversions
and definitions. `subset_config` and `load_dataset` are applied to every study
in the remake.yml file
}
\examples{
\dontrun{
subset_config("data/Falster_2003/metadata.yml", read_yaml("config/traits.yml"),
make_unit_conversion_functions("config/unit_conversions.csv"))
}
}
