% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process.R
\name{substitutions_from_csv}
\alias{substitutions_from_csv}
\title{Substitutions from csv}
\usage{
substitutions_from_csv(
  dataframe_of_substitutions,
  dataset_id,
  trait_name,
  find,
  replace
)
}
\arguments{
\item{dataframe_of_substitutions}{dataframe with columns indicating dataset_id, trait_name, original trait values (find), and AusTraits aligned trait value (replace)}

\item{dataset_id}{study's dataset_id in AusTraits}

\item{trait_name}{trait name for which a trait value replacement needs to be made}

\item{find}{trait value submitted by the contributor for a data observation}

\item{replace}{AusTraits aligned trait value}
}
\value{
modified metadata files with trait value replacements
}
\description{
Function that simultaneously adds many trait value replacements, potentially across many trait_names and dataset_ids, to the respective metadata.yml files.
This function will be used to quickly re-align/re-assign trait values across all AusTraits studies.
}
\examples{
\dontrun{
read_csv("export/dispersal_syndrome_substitutions.csv") \%>\% select(-extra) \%>\%
filter(dataset_id == "Angevin_2011") -> dataframe_of_substitutions
substitutions_from_csv(dataframe_of_substitutions,dataset_id,trait_name,find,replace)
}
}
