% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{test_data_setup}
\alias{test_data_setup}
\title{Test AusTraits studies have the correct format}
\usage{
test_data_setup(dataset_ids = NULL)
}
\arguments{
\item{dataset_ids}{unique study identifier for austraits}
}
\description{
Run the tests to ensure that all compiled studies have the correct format
}
