% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{update_taxonomy}
\alias{update_taxonomy}
\title{Apply taxonomic updates to austraits_raw}
\usage{
update_taxonomy(austraits_raw, taxa)
}
\arguments{
\item{austraits_raw}{AusTraits compiled data as a large list without taxonomic updates applied}

\item{taxa}{taxon list}
}
\value{
list AusTraits compiled data with taxonomic updates applied
}
\description{
Applies taxonomic updates to austraits_raw
}
