% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{user_select_names}
\alias{user_select_names}
\title{Select variable names by user}
\usage{
user_select_names(title, vars)
}
\arguments{
\item{title}{character string providing the instruction for the user}

\item{vars}{variable names}
}
\description{
`user_select names` is used to prompt the user to select the variables that 
are relevant for compiling the metadata yaml file. It is currently used for
`metadata_add_traits`, `metadata_add_sites` and `metadata_add_contexts`
}
