% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{util_check_all_values_in}
\alias{util_check_all_values_in}
\title{Check values in one vector against values in another vector}
\usage{
util_check_all_values_in(x, y, sep = " ")
}
\arguments{
\item{x}{vector}

\item{y}{vector}

\item{sep}{amount of space separating values to be split, default = " " (a single space)}
}
\value{
vector of logical values
}
\description{
`util_check_all_values_in` checks if values in vector x are in y. Values in x may
contain multiple values separated by `sep` so these are split first using `str_split`.
}
