% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{util_get_SHA}
\alias{util_get_SHA}
\title{Get SHA string from Github repository for latest commit}
\usage{
util_get_SHA(path = ".")
}
\arguments{
\item{path}{root directory where a specified file is located, default file name
is the remake.yml file}
}
\value{
40-digit SHA character string for the latest commit to the repository
}
\description{
Get SHA string for the latest commit on Github for the repository. SHA is the
abbreviated SHA-1 40 digit hexadecimal number which Github uses as the
Commit ID to track changes made to a repo
}
