% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_schema.R
\name{util_get_version}
\alias{util_get_version}
\title{Retrieve version for compilation from definitions}
\usage{
util_get_version(path = "config/traits.yml")
}
\arguments{
\item{path}{path to traits definitions}
}
\value{
a string
}
\description{
Retrieve version for compilation from definitions
}
