% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_utils.R
\name{util_kable_styling_html}
\alias{util_kable_styling_html}
\title{Format table with kable and default styling for html}
\usage{
util_kable_styling_html(...)
}
\arguments{
\item{...}{arguments passed to `kableExtra::kable()`}
}
\description{
Format table with kable and default styling for html
}
