% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{util_list_to_bib}
\alias{util_list_to_bib}
\title{Convert a list of elements into a BibEntry object}
\usage{
util_list_to_bib(ref)
}
\arguments{
\item{ref}{list of elements for a reference}
}
\value{
BibEntry object
}
\description{
Convert a list of elements into a BibEntry object
}
