% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{util_list_to_df1}
\alias{util_list_to_df1}
\title{Convert a list with single entries to dataframe}
\usage{
util_list_to_df1(my_list)
}
\arguments{
\item{my_list}{a list with single entries}
}
\value{
a tibble with two columns
}
\description{
Convert a list with single entries to dataframe
}
\examples{
\dontrun{
util_list_to_df1(as.list(iris)[2])
}
}
