% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{util_replace_null}
\alias{util_replace_null}
\title{Convert NULL values to a different value}
\usage{
util_replace_null(x, val = NA)
}
\arguments{
\item{x}{a vector containing null values}

\item{val}{specify what the null value should be returned as, default is NA}
}
\value{
a vector with null values replaced
}
\description{
util_replace_null converts NULL values in a vector to a different value. Default is
converting NuLL to NA
}
\examples{
\dontrun{
util_replace_null(NULL)
}
}
