% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{util_separate_and_sort}
\alias{util_separate_and_sort}
\title{Split and sort cells with multiple values}
\usage{
util_separate_and_sort(x, sep = " ")
}
\arguments{
\item{x}{an individual cell with multiple values}

\item{sep}{a separator, a whitespace is the default}
}
\value{
a vector of alphabetically sorted records
}
\description{
util_separate_and_sort: For a vector x in which individual cell may have
 multiple values (separated by 'sep'), sort records within each cell alphabetically
}
\examples{
\dontrun{util_separate_and_sort("z y x")}
}
