% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{util_standardise_doi}
\alias{util_standardise_doi}
\title{Standarise doi into form https://doi.org/XXX}
\usage{
util_standardise_doi(doi)
}
\arguments{
\item{doi}{doi of reference to add}
}
\description{
Standarise doi into form https://doi.org/XXX
}
