% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{util_strip_taxon_names}
\alias{util_strip_taxon_names}
\title{Strip scientific names of formatting and abbreviations}
\usage{
util_strip_taxon_names(x)
}
\arguments{
\item{x}{vector of names to clean}
}
\value{
vector of cleaned names
}
\description{
Enables better fuzzy matching of scientific names by 
reducing the number of characters that could differ
}
\examples{
c("Bankisa_serrata", "bankisa  serrata", "Banksia Seratta") \%>\% util_strip_taxon_names()
}
