% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_metadata}
\alias{write_metadata}
\title{Write metadata.yml for a study}
\usage{
write_metadata(data, path, style_code = FALSE)
}
\arguments{
\item{data}{austraits metadata object (a list)}

\item{path}{location where the metadata file is to be written to}

\item{style_code}{should the R code be styled?}
}
\description{
write metadata.yml file with custom R code formatted to allow line breaks
}
\examples{
\dontrun{
f <- "data/Falster_2003/metadata.yml"
data <- read_metadata(f)
write_metadata(data, f)
}
}
