% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_metadata_dataset}
\alias{write_metadata_dataset}
\title{Write the YAML representation of metadata.yml for specified `dataset_id` to
file \code{data/dataset_id/metadata.yml}}
\usage{
write_metadata_dataset(metadata, dataset_id)
}
\arguments{
\item{metadata}{metadata file}

\item{dataset_id}{identifier for a particular study in the AusTraits database}
}
\value{
a yml file
}
\description{
Write the YAML representation of metadata.yml for specified `dataset_id` to
file \code{data/dataset_id/metadata.yml}
}
