% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{write_plaintext}
\alias{write_plaintext}
\title{Export AusTraits version as plain text}
\usage{
write_plaintext(austraits, path)
}
\arguments{
\item{austraits}{AusTraits database object}

\item{path}{pathway to save file}
}
\value{
csv files of tibbles containing traits, locations, contexts, methods, excluded_data,
taxonomic updates, taxa, contributors
}
\description{
Export AusTraits version as plain text
}
