% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process.R
\name{convert_month_range_string_to_binary_worker}
\alias{convert_month_range_string_to_binary_worker}
\title{Convert month range to binary}
\usage{
convert_month_range_string_to_binary_worker(str)
}
\arguments{
\item{str}{character string with abbreviated months (i.e. "Jan", "Feb"). Also accepts other
terms such as (all year) and seasons (e.g. summer). The text can also include a range 
separated with "-" (e.g. "Jan-Jul")}
}
\value{
a numeric vector with 12 elements, e.g. c(1,1,1,0,0,0,0,0,0,0,0,1)
}
\description{
Converts flowering and fruiting month ranges to 12 element character strings of 0 and 1 representing Jan - Dec
e.g. c(1,1,1,1,0,0,0,0,0,0,0,1)
}
