% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_format_contributors}
\alias{process_format_contributors}
\title{Format contributors from list into tibble}
\usage{
process_format_contributors(my_list, dataset_id, schema)
}
\arguments{
\item{my_list}{list of input information}

\item{dataset_id}{XXX}

\item{schema}{XXX}
}
\value{
tibble with details of contributors
}
\description{
Format contributors, read in from the metadata.yml file. Converts from list to tibble.
}
\examples{
\dontrun{
process_format_contributors(read_metadata("data/Falster_2003/metadata.yml")$contributors)
}
}
